+++ awi/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## []
 
 - Added `dimensions` configuration to hide dimensions
+- Added `dimensionIcons` configuration to set recipe viewer icon of each dimension
 
 ## [1.1.1]